ConnectionAttemptInjector: fix tsan failures
