+++ test/cpp/end2end/BUILD
@@ -58,10 +58,10 @@ grpc_cc_library(
 )
 
 grpc_cc_library(
-    name = "connection_delay_injector",
+    name = "connection_attempt_injector",
     testonly = True,
-    srcs = ["connection_delay_injector.cc"],
-    hdrs = ["connection_delay_injector.h"],
+    srcs = ["connection_attempt_injector.cc"],
+    hdrs = ["connection_attempt_injector.h"],
     deps = [
         "//:grpc",
     ],
@@ -490,7 +490,7 @@ grpc_cc_test(
     flaky = True,  # TODO(b/151315347)
     tags = ["no_windows"],  # TODO(jtattermusch): fix test on windows
     deps = [
-        ":connection_delay_injector",
+        ":connection_attempt_injector",
         ":test_service_impl",
         "//:gpr",
         "//:grpc",
