+++ test/cpp/end2end/xds/BUILD
@@ -166,7 +166,7 @@ grpc_cc_test(
         "//:grpc",
         "//:grpc++",
         "//test/core/util:grpc_test_util",
-        "//test/cpp/end2end:connection_delay_injector",
+        "//test/cpp/end2end:connection_attempt_injector",
     ],
 )
 
@@ -191,7 +191,7 @@ grpc_cc_test(
         "//:grpc_resolver_fake",
         "//src/proto/grpc/testing/xds/v3:aggregate_cluster_proto",
         "//test/core/util:grpc_test_util",
-        "//test/cpp/end2end:connection_delay_injector",
+        "//test/cpp/end2end:connection_attempt_injector",
     ],
 )
 
@@ -322,7 +322,7 @@ grpc_cc_test(
         "//src/proto/grpc/testing/xds/v3:listener_proto",
         "//src/proto/grpc/testing/xds/v3:route_proto",
         "//test/core/util:grpc_test_util",
-        "//test/cpp/end2end:connection_delay_injector",
+        "//test/cpp/end2end:connection_attempt_injector",
     ],
 )
